Revert "Revert "Refactor/core reorganize 2" (#80)"
This reverts commit b17ea150db50920c006bf96848949231ff66fa4e.
