+++ dags/roger/core/__init__.py
@@ -1,0 +1,4 @@
+"Core roger modules, now broken out into a submodule"
+
+from roger.core.enums import SchemaType, FileFormat
+from roger.core.bulkload import BulkLoad
