+++ dags/roger/core/enums.py
@@ -1,0 +1,15 @@
+"Enums for Roger"
+
+from enum import Enum
+
+class SchemaType(Enum):
+    """ High level semantic metatdata concepts.
+    Categories are classes in an ontological model like Biolink.
+    Predicates are links between nodes. """
+    CATEGORY = "category"
+    PREDICATE = "predicate"
+
+class FileFormat(Enum):
+    """ File formats this module knows about. """
+    JSON = "json"
+    YAML = "yaml"
