+++ dags/roger/models/__init__.py
@@ -1,0 +1,4 @@
+"Data models for Roger"
+
+from roger.models.kgx import KGXModel
+from roger.models.biolink import BiolinkModel
