+++ tests/integration/conftest.py
@@ -1,6 +1,6 @@
 import os
 
-from roger.core import SchemaType
+from roger.core.enums import SchemaType
 import json
 
 class BiolinkMock:
@@ -20,66 +20,57 @@ class BiolinkMock:
                 return y
         return class_names[0]
 
+    def find_biolink_leaves(self, biolink_concepts):
+        return set([concept for concept in biolink_concepts
+                    if concept in ['named_thing', 'some_other_type']])
 
-class UtilMock:
-    category = None
-    predicates = None
-    file_content_assertions = {}
-    kgx_files = []
-    merged_files = []
-    merge_file_test_dir = ''
-    schema = {
-        SchemaType.PREDICATE: {},
-        SchemaType.CATEGORY: {}
-    }
+category = None
+predicates = None
+file_content_assertions = {}
+kgx_files = []
+merged_files = []
+merge_file_test_dir = ''
+schema = {
+    SchemaType.PREDICATE: {},
+    SchemaType.CATEGORY: {}
+}
 
-    def __init__(self):
-        pass
+def kgx_objects():
+    return [os.path.join(*os.path.split(__file__)[:-1], 'data', file)
+            for file in kgx_files]
 
-    @staticmethod
-    def kgx_objects():
-        return [os.path.join(*os.path.split(__file__)[:-1], 'data', file) for file in UtilMock.kgx_files]
+def merged_objects():
+    return [os.path.join(*os.path.split(__file__)[:-1], 'data', file)
+            for file in merged_files]
 
-    @staticmethod
-    def merged_objects():
-        return [os.path.join(*os.path.split(__file__)[:-1], 'data', file) for file in UtilMock.merged_files]
+def bulk_path(*args, **kwargs):
+    return os.path.join(*os.path.split(__file__)[:-1], 'data', 'bulk')
 
-    @staticmethod
-    def bulk_path(*args, **kwargs):
-        return os.path.join(*os.path.split(__file__)[:-1], 'data', 'bulk')
+def is_up_to_date(*args, **kwargs):
+    return False
 
-    @staticmethod
-    def is_up_to_date(*args, **kwargs):
-        return False
+def schema_path(name, *args, **kwargs):
+    return name
 
-    @staticmethod
-    def schema_path(name, *args, **kwargs):
-        return name
+def read_schema(schema_type: SchemaType, *args, **kwargs):
+    return conftest.schema[schema_type]
 
-    @staticmethod
-    def read_schema(schema_type: SchemaType, *args, **kwargs):
-        return UtilMock.schema[schema_type]
+def read_object(path, *args, **kwargs):
+    import json
+    with open(path) as f:
+        return json.load(f)
 
-    @staticmethod
-    def read_object(path, *args, **kwargs):
-        import json
-        with open(path) as f:
-            return json.load(f)
+def write_object(dictionary, file_name):
+    print(dictionary, file_name)
+    print(file_content_assertions)
+    assert file_content_assertions[file_name] == dictionary
 
-    @staticmethod
-    def write_object(dictionary, file_name):
-        print(dictionary, file_name)
-        print(UtilMock.file_content_assertions)
-        assert UtilMock.file_content_assertions[file_name] == dictionary
+def merge_path(file_name):
+    return os.path.join(*os.path.split(__file__)[:-1], 'data', 'merge',
+                        merge_file_test_dir, file_name)
 
-    @staticmethod
-    def merge_path(file_name):
-        return os.path.join(*os.path.split(__file__)[:-1], 'data', 'merge', UtilMock.merge_file_test_dir, file_name)
-
-
-    @staticmethod
-    def json_line_iter(jsonl_file_path):
-        f = open(file=jsonl_file_path, mode='r')
-        for line in f:
-            yield json.loads(line)
-        f.close()
+def json_line_iter(jsonl_file_path):
+    f = open(file=jsonl_file_path, mode='r')
+    for line in f:
+        yield json.loads(line)
+    f.close()
